+++ shuffle/shuffle_test.go
@@ -15,26 +15,29 @@ func TestGetItem(t *testing.T) {
 
 	incoming <- msg
 	time.Sleep(time.Millisecond)
-	item, err := getItem()
-	assert.NoError(err)
+	item := getItem()
 	assert.Equal(msg, item)
 }
 
 func TestFullPool(t *testing.T) {
 	assert := assert.New(t)
 
-	incoming, _ := NewStreamShuffler(5, 1)
+	incoming, _ := NewStreamShuffler(5, 2)
 
 	incoming <- 1
 	incoming <- 2
 	incoming <- 3
 	incoming <- 4
 	incoming <- 5
-	// pool should be full now and one for the buffer
+	// pool should be full. now and one for the buffer
 	incoming <- 6
+	// one for transition
+	incoming <- 7
+	incoming <- 8
+	// buffer should now be full.
 
 	select {
-	case incoming <- 7:
+	case incoming <- 9:
 		assert.Fail("Device Pool should be filled")
 	default:
 	}